Renames app, fixes event
closes #48
